修改

+++ ivan-parent/ivan-core/src/main/java/org/ivan/core/redis/RedisDBUtil.java
@@ -220,7 +220,7 @@ public class RedisDBUtil {
 		// 控制一个pool最多有多少个状态为idle(空闲的)的jedis实例。
 		config.setMaxIdle(ReadPro.getValue("redis.maxIdle", 10000));
 		// 表示当borrow(引入)一个jedis实例时，最大的等待时间，如果超过等待时间，则直接抛出JedisConnectionException；
-		config.setMaxWaitMillis(ReadPro.getValue("redis.maxWait", 1000 * 5));
+		config.setMaxWaitMillis((long) ReadPro.getValue("redis.maxWait", 1000 * 5));
 		// 在borrow一个jedis实例时，是否提前进行validate操作；如果为true，则得到的jedis实例均是可用的；
 		config.setTestOnBorrow(ReadPro.getValue("redis.testOnBorrow", true));
 		// 在return一个jedis实例时，是否提前进行validate操作
